New files:
Lots of icons

activity_event_details_organizer.xml: The UI for the event details page.

EventDetailsOrganizerActivity: Details for events created by the organizer. Editing and deletion functionality work, but need to figure out a way to automatically reload the page. For some reason the drawer isn't working even though it is supposed to be the same as all the other pages.

Edits:

Added Profile functionality to the organizer so that's where a lot fot he small changes come from. Same with adding new icons to the drawer